requirements.txt
puremagic==1.30
pillow==12.0.0
pillow-heif==1.1.1
piexif==1.1.3
PyPDF2==3.0.1
python-docx==1.2.0